refactor: movimentação dos diferentes tipos de logos para uma pasta dedicada
