--- frontend/src/components/Logo.tsx
@@ -1,47 +1,0 @@
-import Svg, { Path } from "react-native-svg";
-import { StyleProp, View, ViewStyle } from "react-native";
-
-interface LogoProps {
-  style?: StyleProp<ViewStyle>;
-}
-
-export function LogoSVG({ style }: LogoProps) {
-  return (
-    <View style={style}>
-      <Svg width={164} height={164} viewBox="0 0 164 164" fill="none">
-        <Path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M141.475 11.6884L142.377 11.6801C142.712 8.31955 143.314 6.90593 145.327 4.6048C145.516 4.38856 145.632 4.34672 145.832 4.1572C146.21 3.79888 146.415 3.64085 146.858 3.34155C147.188 3.1191 147.707 2.78936 148.079 2.6568C146.888 3.42275 145.934 4.30882 145.08 5.53789C144.582 6.25432 143.372 8.74974 143.6 9.79053C147.173 10.0739 149.289 8.58466 150.472 6.63129C151.492 4.94551 151.873 3.14686 152.64 1.50624C153.127 0.465443 153.37 0.44763 153.623 0.0292417C147.974 -0.130036 141.827 1.15454 141.709 6.61907C141.684 7.74457 142.041 8.92435 141.985 9.24808C141.849 10.0242 141.515 10.8218 141.475 11.6884ZM140.852 7.90758C140.685 5.71373 138.698 2.93994 137.333 2.27963C138.696 2.82188 140.618 4.97762 140.998 6.19819C141.197 5.29203 141.182 4.82745 141.632 3.936C141.803 3.59694 141.91 3.60688 141.732 3.25145C140.355 0.49879 135.847 -0.159447 132.637 0.0302773C132.763 0.274475 133.036 0.442452 133.396 1.18954C134.793 4.08305 134.855 8.1642 140.852 7.90758Z"
-          fill="#135961"
-        />
-        <Path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M94.5648 14.7238C95.3127 14.7238 96.0492 14.7675 96.7709 14.8524H136.746C137.467 14.7675 138.204 14.7238 138.952 14.7238C148.247 14.7238 155.782 21.4519 155.782 29.7514C155.782 38.0508 148.247 44.779 138.952 44.779C138.204 44.779 137.467 44.735 136.746 44.6503H96.7709C96.0492 44.735 95.3127 44.779 94.5648 44.779C85.2694 44.779 77.734 38.0508 77.734 29.7514C77.734 21.4519 85.2694 14.7238 94.5648 14.7238Z"
-          fill="#3AA1A8"
-          stroke="#135961"
-          strokeWidth={3.77953}
-          strokeMiterlimit={22.9256}
-        />
-        <Path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M138.952 14.724C148.247 14.724 155.782 21.4522 155.782 29.751C155.782 38.0506 148.247 44.7788 138.952 44.7788C129.656 44.7788 122.121 38.0506 122.121 29.751C122.121 21.4522 129.656 14.724 138.952 14.724Z"
-          fill="#FFAE31"
-          stroke="#135961"
-          strokeWidth={3.77953}
-          strokeMiterlimit={22.9256}
-        />
-        <Path
-          d="M0.000244141 150.803V83.9423H6.24079V145.876H29.3648V150.803H0.000244141ZM36.9306 150.803V97.9674H70.8938V102.894H43.1712V126.637H69.1202V131.564H43.1712V145.876H72.0103V150.803H36.9306ZM94.0172 164L75.8425 108.805H82.5867L94.8054 148.72C95.5711 151.222 96.6125 153.452 97.269 155.593C97.992 153.296 99.0512 151.072 99.7984 148.72L112.477 108.805H118.827L100.488 164H94.0172ZM124.565 150.803V108.805H158.528V113.732H130.805V119.061H156.755V123.988H130.805V145.876H159.645V150.803H124.565Z"
-          fill="#3AA1A8"
-        />
-        <Path
-          d="M29.0117 93.479L10.8371 51.4813H17.5813L29.8 82.0017C30.7745 84.4458 31.6071 86.7334 32.2635 88.8747C32.9866 86.5775 33.7493 84.2616 34.793 82.0017L44.8479 60.2262H51.1977L35.4826 93.479H29.0117ZM60.4789 93.479V29.8577H66.7195V93.479H60.4789ZM77.0295 93.479V51.4813H93.8835C97.5404 51.4813 100.333 51.6789 102.259 52.0778C104.952 52.6356 107.252 53.6422 109.157 55.0888C111.642 56.9661 113.493 59.3514 114.719 62.2645C115.956 65.1775 116.569 68.5112 116.569 72.2557C116.569 75.4425 116.153 78.2677 115.322 80.7312C114.489 83.1947 113.416 85.2376 112.113 86.8511C110.8 88.4639 109.376 89.7344 107.821 90.6539C106.278 91.5822 104.406 92.2881 102.215 92.7655C100.016 93.2448 97.4954 93.479 94.6502 93.479H77.0295ZM83.9203 88.552H93.9496C97.0476 88.552 99.4785 88.2976 101.241 87.7794C103.004 87.2612 104.406 86.5381 105.456 85.5992C106.923 84.2797 108.073 82.5005 108.894 80.2718C109.716 78.0426 110.132 75.3448 110.132 72.1677C110.132 67.7686 109.322 64.3857 107.712 62.0199C106.102 59.6542 104.143 58.0705 101.833 57.2689C100.168 56.6925 97.4854 56.4083 93.7852 56.4083H83.9203V88.552ZM112.337 105.273L138.007 51.4813H144.851L164 93.479H156.916L151.474 80.7505H131.668L118.94 105.273H112.337ZM133.496 76.2344H149.525L144.62 64.5916C143.142 61.052 142.026 58.1293 141.248 55.8224C140.603 58.5595 139.737 61.2476 138.654 63.907L133.496 76.2344Z"
-          fill="#FFAE31"
-        />
-      </Svg>
-    </View>
-  );
-}
