+++ frontend/src/components/logos/BigLogoSVG.tsx
@@ -1,0 +1,52 @@
+import Svg, { Path, G, Defs, ClipPath, Rect } from "react-native-svg";
+import { StyleProp, View, ViewStyle } from "react-native";
+
+interface LogoProps {
+  style?: StyleProp<ViewStyle>;
+}
+
+export function BigLogoSVG(props: LogoProps) {
+  return (
+    <Svg width={221} height={221} viewBox="0 0 221 221" fill="none" {...props}>
+      <G clipPath="url(#clip0_3033_10896)">
+        <Path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M190.517 15.846L191.732 15.8348C192.182 11.3062 192.994 9.40129 195.707 6.30037C195.962 6.00898 196.117 5.95259 196.387 5.69721C196.896 5.21435 197.173 5.00139 197.77 4.59807C198.214 4.2983 198.915 3.85396 199.415 3.67533C197.811 4.70748 196.525 5.90152 195.374 7.55776C194.703 8.52321 193.072 11.8859 193.379 13.2885C198.195 13.6703 201.046 11.6635 202.639 9.03119C204.015 6.7595 204.528 4.33571 205.562 2.12487C206.218 0.722336 206.545 0.698332 206.886 0.134528C199.274 -0.080108 190.99 1.65094 190.831 9.01472C190.798 10.5314 191.279 12.1212 191.203 12.5575C191.02 13.6033 190.57 14.6782 190.517 15.846ZM189.677 10.7511C189.452 7.79473 186.774 4.05687 184.934 3.16707C186.771 3.89778 189.362 6.80277 189.873 8.44757C190.141 7.22646 190.121 6.60041 190.727 5.39912C190.958 4.94221 191.102 4.95561 190.862 4.47666C189.006 0.767273 182.933 -0.119742 178.606 0.135924C178.776 0.464996 179.144 0.691355 179.629 1.69811C181.512 5.59729 181.595 11.0969 189.677 10.7511Z"
+          fill="#135961"
+        />
+        <Path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M127.302 19.9363C128.309 19.9363 129.302 19.9952 130.274 20.1097H184.143C185.115 19.9952 186.108 19.9363 187.116 19.9363C199.642 19.9363 209.796 29.003 209.796 40.1869C209.796 51.3709 199.642 60.4375 187.116 60.4375C186.108 60.4375 185.115 60.3784 184.143 60.2642H130.274C129.302 60.3784 128.309 60.4375 127.302 60.4375C114.776 60.4375 104.621 51.3709 104.621 40.1869C104.621 29.003 114.776 19.9363 127.302 19.9363Z"
+          fill="#3AA1A8"
+          stroke="#135961"
+          strokeWidth={3.77953}
+          strokeMiterlimit={22.9256}
+        />
+        <Path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M187.115 19.9366C199.641 19.9366 209.796 29.0032 209.796 40.1864C209.796 51.3706 199.641 60.4373 187.115 60.4373C174.59 60.4373 164.435 51.3706 164.435 40.1864C164.435 29.0032 174.59 19.9366 187.115 19.9366Z"
+          fill="#FFAE31"
+          stroke="#135961"
+          strokeWidth={3.77953}
+          strokeMiterlimit={22.9256}
+        />
+        <Path
+          d="M-0.129883 203.311V113.212H8.27963V196.672H39.4407V203.311H-0.129883ZM49.6361 203.311V132.112H95.4035V138.752H58.0456V170.746H93.0134V177.385H58.0456V196.672H96.9081V203.311H49.6361ZM126.564 221.095L102.072 146.717H111.16L127.626 200.505C128.658 203.876 130.061 206.881 130.946 209.767C131.92 206.671 133.347 203.674 134.354 200.505L151.44 146.717H159.997L135.284 221.095H126.564ZM167.728 203.311V146.717H213.496V153.356H176.138V160.537H211.106V167.177H176.138V196.672H215.001V203.311H167.728Z"
+          fill="#3AA1A8"
+        />
+        <Path
+          d="M38.9651 126.064L14.4736 69.4693H23.5618L40.0273 110.597C41.3405 113.891 42.4624 116.974 43.3471 119.859C44.3214 116.764 45.3493 113.643 46.7556 110.597L60.3052 81.2536H68.8619L47.685 126.064H38.9651ZM81.369 126.064V40.3301H89.7785V126.064H81.369ZM103.672 126.064V69.4693H126.384C131.312 69.4693 135.075 69.7356 137.671 70.2731C141.299 71.0248 144.399 72.3813 146.966 74.3306C150.315 76.8604 152.809 80.0747 154.46 84.0004C156.128 87.9258 156.954 92.4181 156.954 97.4641C156.954 101.759 156.394 105.566 155.273 108.885C154.151 112.205 152.705 114.958 150.95 117.132C149.179 119.306 147.261 121.018 145.165 122.257C143.086 123.508 140.563 124.459 137.611 125.102C134.648 125.748 131.251 126.064 127.417 126.064H103.672ZM112.958 119.424H126.473C130.648 119.424 133.923 119.082 136.299 118.383C138.674 117.685 140.563 116.71 141.979 115.445C143.956 113.667 145.505 111.269 146.611 108.266C147.719 105.262 148.279 101.627 148.279 97.3455C148.279 91.4175 147.187 86.8587 145.018 83.6707C142.85 80.4827 140.208 78.3486 137.096 77.2685C134.853 76.4917 131.237 76.1088 126.251 76.1088H112.958V119.424ZM151.252 141.957L185.843 69.4693H195.065L220.87 126.064H211.324L203.991 108.911H177.301L160.148 141.957H151.252ZM179.764 102.826H201.364L194.755 87.1362C192.763 82.3665 191.258 78.4279 190.21 75.3192C189.341 79.0076 188.175 82.6299 186.714 86.2137L179.764 102.826Z"
+          fill="#FFAE31"
+        />
+      </G>
+      <Defs>
+        <ClipPath id="clip0_3033_10896">
+          <Rect width={221} height={221} fill="white" />
+        </ClipPath>
+      </Defs>
+    </Svg>
+  );
+}
