+++ frontend/src/components/logos/HorizontalLogoSVG.tsx
